Forgot to add bugfix lmao

+++ Runtime/Core/Src/RObject/Names.h
@@ -24,7 +24,7 @@ struct RNameID
 
         RNameID(RNameIDList Input)
         {
-            Var = Input;
+            Var = static_cast<uint32>(Input);
         }
 
         uint32 Get()
